use doubles for amplitudes and soft clip them

+++ src/instr.h
@@ -1,3 +1,3 @@
 /* #include global.h */
 
-int instr_get(Instr* instr, int note_on, long phase, long srate);
+double instr_get(Instr* instr, int note_on, long phase, long srate);
